this is a mess, sorry

but

before:
```
$ .build/release/bench
Benchmark result:
Took 7.7 seconds to run
648582.28 req/sec
```

after:
```
$ .build/release/bench
Benchmark result:
Took 5.27 seconds to run
948171.8 req/sec
```